+++ global.d.ts
@@ -1,0 +1,6 @@
+declare module 'tti-polyfill' {
+  const ttiPolyfill: {
+    getFirstConsistentlyInteractive(): Promise<number>;
+  };
+  export default ttiPolyfill;
+}